ni-35 support configuration
