+++ inio.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    pattern: "./terms/<key>.json",
+};
